requirements.txt
flask==3.1.2
flask-cors==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==2.4.0
joblib==1.3.1
requests==2.31.0
python-whois==0.9.3